Create FeedItem component
